+++ pfl/odtable.h
@@ -160,7 +160,7 @@ void	 pfl_odt_allocitem(struct pfl_odt *, void **);
 size_t	 pfl_odt_allocslot(struct pfl_odt *);
 void	 pfl_odt_check(struct pfl_odt *, void (*)(void *, int64_t, void *), 
 	    void *);
-void	 pfl_odt_create(const char *, size_t, size_t, int, size_t,
+void	 pfl_odt_create(const char *, int64_t, size_t, int, size_t,
 	    size_t, int);
 void	 pfl_odt_freeitem(struct pfl_odt *, int64_t);
 void	 pfl_odt_getslot(struct pfl_odt *,
